ci: exercise Transformers compatibility floor (#478)
Pin the lowest resolvable Transformers and declared TRL versions in a dedicated CI job, guard production model-loading kwargs, and retain compatibility with pre-4.56 Transformers.

+++ src/soup_cli/commands/merge.py
@@ -175,7 +175,7 @@ def merge(
         console.print(f"[dim]Loading base model: {base}...[/]")
         model = AutoModelForCausalLM.from_pretrained(
             base,
-            dtype=model_dtype,
+            torch_dtype=model_dtype,
             trust_remote_code=trc,
             device_map="cpu",
         )
